requirements.txt
Django<3.3
graphene-django
djangorestframework-simplejwt
